--- src/shared/api/adapters/axios/AxiosInstance.ts
@@ -1,6 +1,0 @@
-import axios from "axios";
-import { API } from "src/app.constants";
-
-const AxiosInstance = axios.create({ baseURL: API.BASE_URL });
-
-export { AxiosInstance };
